+++ src/DI/Definition/Resolver/ResolverDispatcher.php
@@ -3,7 +3,7 @@
 namespace DI\Definition\Resolver;
 
 use DI\Definition\Definition;
-use DI\Definition\Exception\DefinitionException;
+use DI\Definition\Exception\InvalidDefinition;
 use DI\Proxy\ProxyFactory;
 use Psr\Container\ContainerInterface;
 
@@ -47,7 +47,7 @@ class ResolverDispatcher implements DefinitionResolver
      * @param Definition $definition Object that defines how the value should be obtained.
      * @param array      $parameters Optional parameters to use to build the entry.
      *
-     * @throws DefinitionException If the definition cannot be resolved.
+     * @throws InvalidDefinition If the definition cannot be resolved.
      *
      * @return mixed Value obtained from the definition.
      */