+++ tests/UnitTest/Definition/Source/Fixtures/definitions-fail.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'foo' => 'bar',
+    'baz', // error => this entry is not indexed by a string
+];